+++ assets/ds4a_styles.css
@@ -5,17 +5,24 @@
 }
 
 .ds4a-body {
-    margin-left: 18rem;
+    margin-left: 2rem;
     margin-right: 2rem;
-    padding: 2rem 1rem;
-    background-color: #F8F9F9;
+    margin-top: 2rem;
+    padding: 5rem 5rem;
+    background-color: #878C8C;
 }
 
+.title_wrap {
+    width: 100%;
+    background-color: #878C8C;
+    position: fixed;
+}
+
+/*
 .ds4a-title {
-    position: relative;
+    position: fixed;
     top: 0;
     left: 0;
-    z-index: 999;
     height: 80px;
     width: auto;
     margin-left: 16rem;
@@ -24,9 +31,22 @@
     padding: 1rem 1rem;
     background-color: #878C8C;
 }
+ */
+.ds4a-title {
+    position: relative;
+    top: 0;
+    left: 0;
+    height: 60px;
+    width: 100%;
+    margin-left: 12rem;
+    margin-right: 12rem;
+    margin-bottom: 0px;
+    padding: 1rem 1rem;
+    background-color: #878C8C;
+}
 
 .ds4a-sidebar {
-    position: relative;
+    position: fixed;
     top: 50;
     left: 0;
     bottom: 0;
@@ -41,7 +61,7 @@
 }
 .custom-tabs {
     border-top-left-radius: 5px;
-    background-color: #F8F9F9;
+    background-color: #878C8C;
     padding: 0px 24px;
     border-bottom: 1px solid #d6d6d6;
 }
@@ -89,9 +109,4 @@
 
 .flex-display {
   display: flex;
-}
-
-#right-column {
-  display: flex;
-  flex-direction: column;
 }
